Move OngoingOperationTracker to osu.Game.Online namespace
