+++ osu.Game/Online/OngoingOperationTracker.cs
@@ -1,4 +1,4 @@
-// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
 using System;
@@ -6,7 +6,7 @@ using osu.Framework.Allocation;
 using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 
-namespace osu.Game.Screens.OnlinePlay
+namespace osu.Game.Online
 {
     /// <summary>
     /// Utility class to track ongoing online operations' progress.
